Started adding errors, added eulers step method

+++ NumericalMethods/MidpointRule.cs
@@ -4,11 +4,11 @@ namespace NumericalMethods
 {
     class MidpointRule
     {
-        public static double Calculate(long n, double start, double end, Func<double, double> f)
+        public static decimal Calculate(long n, decimal start, decimal end, Func<decimal, decimal> f)
         {
-            double interval = (end - start) / n;
-            double total = 0;
-            double current = start;
+            decimal interval = (end - start) / n;
+            decimal total = 0;
+            decimal current = start;
             for (long i = 0; i < n; i++)
                 total += f((current + (current += interval)) / 2);
             return interval * total;
