+++ NumericalMethods/TrapeziumRule.cs
@@ -4,17 +4,17 @@ namespace NumericalMethods
 {
     class TrapeziumRule
     {
-        public static double Calculate(long n, double start, double end, Func<double, double> f)
+        public static decimal Calculate(long n, decimal start, decimal end, Func<decimal, decimal> f)
         {
-            double interval = (end - start) / n;
-            double total = 0;
-            double current = start;
+            decimal interval = (end - start) / n;
+            decimal total = 0;
+            decimal current = start;
             for (long i = 0; i < n - 1; i++)
                 total += f((current += interval));
             return (interval * (f(start) + f(end) + (2 * total))) / 2;
         }
 
-        public static double Calculate(double mn, double tn)
+        public static decimal Calculate(decimal mn, decimal tn)
         {
             return (mn + tn) / 2;
         }
